Added Alert Tables for Landing page, this includes quantity alerts and expiration alerts

+++ src/main/java/com/storagemanager/storagemanager/productBatch/ProductBatchRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ProductBatchRepository extends CrudRepository<ProductBatchEntry, Long> {
 
     List<ProductBatchEntry> findAll();
-    List<ProductBatchEntry> findByproductSku(int productSku);
+    List<ProductBatchEntry> findByproductSku(Long productSku);
     List<ProductBatchEntry> findBybatchId(Long batchId);
 
 }
